[Java 17]: Drop Nashorn
This changeset drops Nashorn-specific classes which are no longer
available in JDK15+ in accordance with JEP372[1], and replaces
them with JSR223[2]-compatible calls.

These calls work with GraalVM[3] to support a seamless switch to
Graal JS when running on that platform. For other JDK17 runtimes,
Graal JS dependencies can be included directly to run on a stock
stock JVM[4].

Changes enclosed:
- Drop references to Nashorn classes, replace with JSR223 calls
- Configure GraalJS with predicate-based class filter
- Fixup references to script mirrors as maps
- Compatibility with pre-JDK17 via Maven deps

[1]: https://openjdk.org/jeps/372
[2]: https://jcp.org/en/jsr/detail?id=223
[3]: https://www.graalvm.org/22.1/reference-manual/js/NashornMigrationGuide/
[4]: https://www.graalvm.org/22.1/reference-manual/js/RunOnJDK/